+++ src/ts/components/recommendationoverlay.ts
@@ -57,7 +57,7 @@ export class RecommendationOverlay extends Container<ContainerConfig> {
     };
 
     // Add recommendation when a source is loaded
-    player.on(player.exports.Event.Ready, setupRecommendations);
+    player.on(player.exports.Event.SourceLoaded, setupRecommendations);
     // Remove recommendations and hide overlay when source is unloaded
     player.on(player.exports.Event.SourceUnloaded, () => {
       clearRecommendations();
@@ -65,12 +65,6 @@ export class RecommendationOverlay extends Container<ContainerConfig> {
     });
     // Display recommendations when playback has finished
     player.on(player.exports.Event.PlaybackFinished, () => {
-      // Dismiss ON_PLAYBACK_FINISHED events at the end of ads
-      // TODO remove this workaround once issue #1278 is solved
-      if (player.isAd()) {
-        return;
-      }
-
       this.show();
     });
     // Hide recommendations when playback starts, e.g. a restart