+++ src/ts/main.ts
@@ -73,6 +73,7 @@ import {SettingsPanelPageBackButton} from './components/settingspanelpagebackbut
 import {SettingsPanelPageOpenButton} from './components/settingspanelpageopenbutton';
 import {SubtitleSettingsPanelPage} from './components/subtitlesettings/subtitlesettingspanelpage';
 import {SettingsPanelItem} from './components/settingspanelitem';
+import {ErrorUtils} from './errorutils';
 
 // Object.assign polyfill for ES5/IE9
 // https://developer.mozilla.org/de/docs/Web/JavaScript/Reference/Global_Objects/Object/assign
@@ -110,6 +111,7 @@ const playerui = {
   UIUtils,
   BrowserUtils,
   StorageUtils,
+  ErrorUtils,
   // Components
   AdClickOverlay,
   AdMessageLabel,